+++ demisto_sdk/commands/validate/tests/validators_test.py
@@ -41,13 +41,13 @@ from demisto_sdk.tests.constants_test import (
     CONF_JSON_MOCK_PATH, DASHBOARD_TARGET, DIR_LIST, IGNORED_PNG,
     INCIDENT_FIELD_TARGET, INCIDENT_TYPE_TARGET, INDICATOR_TYPE_TARGET,
     INTEGRATION_RELEASE_NOTES_TARGET, INTEGRATION_TARGET,
-    INVALID_DASHBOARD_PATH, INVALID_IGNORED_UNIFIED_INTEGRATION,
-    INVALID_INCIDENT_FIELD_PATH, INVALID_INTEGRATION_ID_PATH,
-    INVALID_INTEGRATION_NO_TESTS, INVALID_INTEGRATION_NON_CONFIGURED_TESTS,
-    INVALID_LAYOUT_CONTAINER_PATH, INVALID_LAYOUT_PATH,
-    INVALID_MULTI_LINE_1_CHANGELOG_PATH, INVALID_MULTI_LINE_2_CHANGELOG_PATH,
-    INVALID_ONE_LINE_1_CHANGELOG_PATH, INVALID_ONE_LINE_2_CHANGELOG_PATH,
-    INVALID_ONE_LINE_LIST_1_CHANGELOG_PATH,
+    INVALID_BETA_INTEGRATION, INVALID_DASHBOARD_PATH,
+    INVALID_IGNORED_UNIFIED_INTEGRATION, INVALID_INCIDENT_FIELD_PATH,
+    INVALID_INTEGRATION_ID_PATH, INVALID_INTEGRATION_NO_TESTS,
+    INVALID_INTEGRATION_NON_CONFIGURED_TESTS, INVALID_LAYOUT_CONTAINER_PATH,
+    INVALID_LAYOUT_PATH, INVALID_MULTI_LINE_1_CHANGELOG_PATH,
+    INVALID_MULTI_LINE_2_CHANGELOG_PATH, INVALID_ONE_LINE_1_CHANGELOG_PATH,
+    INVALID_ONE_LINE_2_CHANGELOG_PATH, INVALID_ONE_LINE_LIST_1_CHANGELOG_PATH,
     INVALID_ONE_LINE_LIST_2_CHANGELOG_PATH, INVALID_PLAYBOOK_CONDITION_1,
     INVALID_PLAYBOOK_CONDITION_2, INVALID_PLAYBOOK_ID_PATH,
     INVALID_PLAYBOOK_PATH, INVALID_PLAYBOOK_PATH_FROM_ROOT,
@@ -380,6 +380,64 @@ class TestValidators:
         """
         validate_manager = ValidateManager(file_path=file_path, skip_conf_json=True)
         assert validate_manager.run_validation_on_specific_files()
+
+    INVALID_FILES_PATHS_FOR_ALL_VALIDATIONS = [
+        INVALID_DASHBOARD_PATH,
+        INVALID_INCIDENT_FIELD_PATH,
+        INVALID_INTEGRATION_ID_PATH,
+        INVALID_INTEGRATION_NO_TESTS,
+        INVALID_INTEGRATION_NON_CONFIGURED_TESTS,
+        INVALID_LAYOUT_CONTAINER_PATH,
+        INVALID_LAYOUT_PATH,
+        INVALID_PLAYBOOK_CONDITION_1,
+        INVALID_PLAYBOOK_CONDITION_2,
+        INVALID_PLAYBOOK_ID_PATH,
+        INVALID_PLAYBOOK_PATH,
+        INVALID_PLAYBOOK_PATH_FROM_ROOT,
+        INVALID_REPUTATION_PATH,
+        INVALID_SCRIPT_PATH,
+        INVALID_WIDGET_PATH,
+        INVALID_BETA_INTEGRATION
+    ]
+
+    @pytest.mark.parametrize('file_path', INVALID_FILES_PATHS_FOR_ALL_VALIDATIONS)
+    @patch.object(ImageValidator, 'is_valid', return_value=True)
+    def test_run_all_validations_on_file_failed(self, _, file_path):
+        """
+        Given
+        - An invalid file inside a pack
+
+        When
+        - running run_all_validations_on_file on that file
+
+        Then
+        -  The file will be validated and failed
+        """
+        validate_manager = ValidateManager(file_path=file_path, skip_conf_json=True)
+        assert not validate_manager.run_validation_on_specific_files()
+
+    def test_run_all_validations_on_file_white_modified_id(self, mocker):
+        """
+        Given
+        - Integration with a modified ID.
+
+        When
+        - running 'run_all_validations_on_file' on that file.
+
+        Then
+        -  The file will fail validation because its id changed.
+        """
+
+        validator = StructureValidator(file_path=VALID_INTEGRATION_ID_PATH)
+        old = validator.load_data_from_file()
+        old['commonfields']['id'] = 'old_id'
+
+        mocker.patch.object(ImageValidator, 'is_valid', return_value=True)
+        mocker.patch('demisto_sdk.commands.common.hook_validations.structure.is_file_path_in_pack', return_value=True)
+        mocker.patch('demisto_sdk.commands.common.hook_validations.structure.get_remote_file', return_value=old)
+
+        validate_manager = ValidateManager(file_path=VALID_INTEGRATION_ID_PATH, skip_conf_json=True)
+        assert not validate_manager.run_validation_on_specific_files()
 
     def test_files_validator_validate_pack_unique_files(self):
         validate_manager = ValidateManager(skip_conf_json=True)