+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fixed an issue in the **format** command where format failed when trying to update invalid Docker image.
 * The **format** command will now preserve the **isArray** argument in integration's reputation commands and will show a warning if it set to **false**.
 * Fixed an issue in the **lint** command where *finally* clause was not supported in main function.
+* Fixed an issue in the **validate** command where changing any entity ID was not validated.
 
 
 # v1.2.8